already exists error handling

+++ src/project.rs
@@ -4,10 +4,10 @@ use std::path::Path;
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct ProjectConfig {
-  name: String,
-  path: Option<String>,
-  editor: Option<String>,
-  terminal: Option<String>,
+  pub name: String,
+  pub path: Option<String>,
+  pub editor: Option<String>,
+  pub terminal: Option<String>,
 }
 
 impl ProjectConfig {
@@ -29,7 +29,12 @@ impl ProjectConfig {
 
   pub fn save_to(&self, path: &Path) -> io::Result<()> {
     let file_name = format!("{}.yml", self.name);
-    let file_path = path.join(file_name);
+    let file_path = path.join(&file_name);
+
+    if file_path.exists() {
+      return Err(io::Error::new(io::ErrorKind::AlreadyExists, format!("File '{}' already exists.", file_name)))
+    }
+
     let contents = serde_yaml::to_string(&self)
       .map_err(|e| io::Error::new(io::ErrorKind::Other, e.to_string()))?;
     fs::write(file_path, contents)?;
